Isolated Location specific queries in Location class. Cleaned and documented.
